Pull ArticleSeries into own component, update loop logic
